+++ Changelog.md
@@ -12,6 +12,7 @@ Bugfixes:
  * ABIEncoderV2: Fix internal error related to bare delegatecall.
  * ABIEncoderV2: Fix internal error related to ecrecover.
  * ABIEncoderV2: Fix internal error related to mappings as library parameters.
+ * SMTChecker: Fixed crash when used with fixed-sized arrays.
  * Yul: Properly detect name clashes with functions before their declaration.